# Refactor: `@truthtable` macro