--- a/src/macroutils.jl
+++ b/src/macroutils.jl
@@ -1,8 +1,46 @@
-function gencolumns(n::Integer)
+function _gencolumns(n::Integer)
   bools = [true, false]
   outers = [2^x for x in 0:n-1]
   inners = reverse(outers)
   [repeat(bools; inner, outer) for (inner, outer) in zip(inners, outers)]
+end
+
+function _colexpr(expr::Expr)
+  colexpr = copy(expr)
+  _preprocess!(colexpr)
+  colexpr
+end
+
+function _preprocess!(expr::Expr)
+  if expr.head === :(-->)
+    expr.head = :call
+    pushfirst!(expr.args, :(-->))
+  end
+
+  if expr.head === :&&
+    expr.head = :call
+    pushfirst!(expr.args, :&)
+  end
+
+  if expr.head === :||
+    expr.head = :call
+    pushfirst!(expr.args, :|)
+  end
+
+  _checkexpr(expr)
+
+  pushfirst!(expr.args, :broadcast)
+
+  for i in 3:length(expr.args)
+    arg = expr.args[i]
+    if arg isa Symbol
+      expr.args[i] = :(colmap[$(QuoteNode(arg))])
+    end
+
+    if arg isa Expr
+      _preprocess!(arg)
+    end
+  end
 end
 
 const OPRS = (
@@ -15,33 +53,27 @@ const OPRS = (
   :(===), :≡
 )
 
-function preprocess!(expr::Expr)
-  if expr.head == :(-->)
-    expr.head = :call
-    expr.args = [:(-->); expr.args]
+function _checkexpr(expr::Expr)
+  if expr.head !== :call
+    throw(ArgumentError("Invalid expression"))
   end
 
-  if expr.head ∈ (:&&, :||)
-    args = expr.args
-  elseif expr.head == :call && expr.args[1] ∈ OPRS
-    args = expr.args[2:end]
-  else
+  if expr.args[1] ∉ OPRS
     throw(ArgumentError("Expression with invalid operator"))
-  end
-
-  for arg in args
-    if arg isa Expr
-      preprocess!(arg)
-    end
   end
 end
 
-_propname(x) = throw(ArgumentError("$x is not a valid proposition name"))
-_propname(x::Symbol) = x
+function _propnames(expr::Expr)
+  props = Symbol[]
+  _propnames!(props, expr)
+  unique!(props)
+  return props
+end
 
 function _propnames!(props::Vector{Symbol}, expr::Expr)
   b = expr.head == :call ? 2 : 1
-  for arg in expr.args[b:end]
+  for i in length(expr.args):-1:b
+    arg = expr.args[i]
     if arg isa Expr
       _propnames!(props, arg)
     else
@@ -50,31 +82,28 @@ function _propnames!(props::Vector{Symbol}, expr::Expr)
   end
 end
 
-function propnames(expr::Expr)
-  props = Symbol[]
-  _propnames!(props, expr)
-  unique!(props)
-  return props
+_propname(x) = throw(ArgumentError("$x is not a valid proposition name"))
+_propname(x::Symbol) = x
+
+function _subexprs(expr::Expr)
+  exprs = [expr]
+  _subexprs!(exprs, expr)
+  return exprs
 end
 
-function _getsubexprs!(exprs::Vector{Expr}, expr::Expr)
+function _subexprs!(exprs::Vector{Expr}, expr::Expr)
   b = expr.head == :call ? 2 : 1
-  for arg in expr.args[end:-1:b]
+  for i in length(expr.args):-1:b
+    arg = expr.args[i]
     if arg isa Expr
       pushfirst!(exprs, arg)
-      _getsubexprs!(exprs, arg)
+      _subexprs!(exprs, arg)
     end
   end
 end
 
-function getsubexprs(expr::Expr)
-  exprs = [expr]
-  _getsubexprs!(exprs, expr)
-  return exprs
-end
-
 @static if VERSION < v"1.7"
-  function exprname(expr::Expr)
+  function _exprname(expr::Expr)
     str = string(expr)
     str = replace(str, r"&{2}|&" => "∧")
     str = replace(str, r"\|{2}|\|" => "∨")
@@ -84,7 +113,7 @@ end
     Symbol(replace(str, "===" => "≡"))
   end
 else
-  function exprname(expr::Expr)
+  function _exprname(expr::Expr)
     str = string(expr)
     str = replace(str,
       r"&{2}|&" => "∧",