--- a/src/truthtable.jl
+++ b/src/truthtable.jl
@@ -4,7 +4,7 @@ struct TruthTable
   colnames::Vector{Symbol}
 end
 
-function TruthTable(columns::Vector{<:AbstractVector{Bool}}, colnames::Vector{Symbol})
+function TruthTable(columns::Vector{Vector{Bool}}, colnames::Vector{Symbol})
   colindex = Dict(k => i for (i, k) in enumerate(colnames))
   TruthTable(columns, colindex, colnames)
 end