This PR is a refactoring of the `@truthtable` macro to make it more efficient.
Benchmarks:
main:
```julia
julia> @time @eval @truthtable !(x || y) <--> (!x && !y) full=true
  0.617676 seconds (2.64 M allocations: 134.338 MiB, 6.41% gc time, 98.37% compilation time)
TruthTable
...

julia> @time @eval @truthtable !(x || y) <--> (!x && !y) full=true
  0.136912 seconds (370.69 k allocations: 18.694 MiB, 96.07% compilation time)
TruthTable
...

julia> @time @eval @truthtable !(x || y) <--> (!x && !y) full=true
  0.141954 seconds (370.68 k allocations: 18.689 MiB, 96.22% compilation time)
TruthTable
...
```
This PR:
```julia
julia> @time @eval @truthtable !(x || y) <--> (!x && !y) full=true
  0.531565 seconds (2.17 M allocations: 111.631 MiB, 5.38% gc time, 98.84% compilation time)
TruthTable
...

julia> @time @eval @truthtable !(x || y) <--> (!x && !y) full=true
  0.001128 seconds (515 allocations: 26.062 KiB, 0.02% compilation time)
TruthTable
...

julia> @time @eval @truthtable !(x || y) <--> (!x && !y) full=true
  0.001776 seconds (512 allocations: 25.969 KiB, 0.03% compilation time)
TruthTable
...
```